+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 # [Release 3.11](https://github.com/GENI-NSF/geni-portal/milestones/3.11)
 
+* Remove Flack
+  ([#1676](https://github.com/GENI-NSF/geni-portal/issues/1676))
+* Support Adding an RSPEC from a File to an existing Jacks topology
+  ([#1660](https://github.com/GENI-NSF/geni-portal/issues/1660))
+
 
 # [Release 3.10](https://github.com/GENI-NSF/geni-portal/milestones/3.10)